itest: move funding tests into one file
